refactor(child-process): add env to command, move builder to separate file
