+++ crates/rmcp/src/transport/child_process2.rs
@@ -1,3 +1,4 @@
+pub mod builder;
 pub mod runner;
 pub mod transport;
